backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt==4.0.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Redis
redis>=5.0.0
celery>=5.3.0

# Utils
python-dateutil>=2.8.0
pytz>=2024.1

# Monitoring
prometheus-client>=0.20.0

# File handling
python-magic>=0.4.27
pillow>=10.2.0
aiofiles>=23.2.1